e2e tests for v1 apis (#2653)
* chore: Update test.use options in e2e tests

* chore: Update Authorization header in e2e tests

* feat: Add getConnectionByProduct function for retrieving SSO connections by product

* feat: Add test for retrieving empty array of SSO connections by product

* feat: Add OAuth helper functions for authorization

* chore: Remove unused import in oauth.ts file

* feat: Add getDirectoryByProduct function for retrieving a directory by product

* lint fixes

* feat: e2e tests

* lint fixes

* feat: Add getDirectoryEvents function for retrieving events from a directory

* feat: Add deleteAllEventsFromDirectory function for deleting all events from a directory

* feat: Refactor SSO trace API functions to remove unnecessary object destructuring

* refactor: Update webhook URL and secret to empty string in directoryPayload

+++ e2e/api/v1/directory-sync/directories.spec.ts
@@ -5,14 +5,11 @@ import {
   directoryExpected,
   directoryPayload,
   getDirectory,
+  getDirectoryByProduct,
 } from '../../helpers/directories';
+import { options } from '../../helpers/api';
 
-test.use({
-  extraHTTPHeaders: {
-    Authorization: `Api-Key secret`,
-    'Content-Type': 'application/json',
-  },
-});
+test.use(options);
 
 const { tenant, product } = directoryPayload;
 
@@ -25,6 +22,9 @@ test.beforeAll(async ({ request }) => {
 test.afterAll(async ({ request }) => {
   const [directory] = await getDirectory(request, { tenant, product });
 
+  if (!directory) {
+    return;
+  }
   await deleteDirectory(request, directory.id);
 });
 
@@ -155,5 +155,17 @@ test.describe('PATCH /api/v1/dsync/{directoryId}', () => {
         },
       },
     });
+  });
+});
+
+test.describe('GET /api/v1/dsync/product', () => {
+  test('should be able to get a directory by product', async ({ request }) => {
+    let directories = await getDirectoryByProduct(request, { product });
+    expect(directories.length).toBe(1);
+
+    await deleteDirectory(request, directories[0].id);
+
+    directories = await getDirectoryByProduct(request, { product });
+    expect(directories.length).toBe(0);
   });
 });
